WIP move out patched methods to separate modules
