+++ ddtrace/contrib/internal/langchain/utils.py
@@ -1,5 +1,24 @@
 import inspect
 import sys
+from typing import Any
+from typing import Optional
+from typing import Union
+
+import langchain
+from pydantic import SecretStr
+
+from ddtrace.internal.utils.version import parse_version
+
+
+def get_version():
+    # type: () -> str
+    return getattr(langchain, "__version__", "")
+
+
+# After 0.1.0, implementation split into langchain, langchain_community, and langchain_core.
+# We need to check the version to determine which module to wrap, to avoid deprecation warnings
+# ref: https://github.com/DataDog/dd-trace-py/issues/8212
+PATCH_LANGCHAIN_V0 = parse_version(get_version()) < (0, 1, 0)
 
 
 class BaseTracedLangChainStreamResponse:
@@ -128,3 +147,36 @@ def tag_general_message_input(span, inputs, integration, langchain_core):
             span.set_tag_str("langchain.request.messages.%d.message_type" % (input_idx), str(role))
         else:
             span.set_tag_str("langchain.request.messages.%d.content" % (input_idx), integration.trunc(str(inp)))
+
+
+def _extract_model_name(instance: Any) -> Optional[str]:
+    """Extract model name or ID from llm instance."""
+    for attr in ("model", "model_name", "model_id", "model_key", "repo_id"):
+        if hasattr(instance, attr):
+            return getattr(instance, attr)
+    return None
+
+
+def _format_api_key(api_key: Union[str, SecretStr]) -> str:
+    """Obfuscate a given LLM provider API key by returning the last four characters."""
+    if hasattr(api_key, "get_secret_value"):
+        api_key = api_key.get_secret_value()
+
+    if not api_key or len(api_key) < 4:
+        return ""
+    return "...%s" % api_key[-4:]
+
+
+def _extract_api_key(instance: Any) -> str:
+    """
+    Extract and format LLM-provider API key from instance.
+    Note that langchain's LLM/ChatModel/Embeddings interfaces do not have a
+    standard attribute name for storing the provider-specific API key, so make a
+    best effort here by checking for attributes that end with `api_key/api_token`.
+    """
+    api_key_attrs = [a for a in dir(instance) if a.endswith(("api_token", "api_key"))]
+    if api_key_attrs and hasattr(instance, str(api_key_attrs[0])):
+        api_key = getattr(instance, api_key_attrs[0], None)
+        if api_key:
+            return _format_api_key(api_key)
+    return ""
